movie_task/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
Django==3.2.7
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
exceptiongroup==1.1.3
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.3.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
PyJWT==1.7.1
pytest==7.4.1
pytest-django==4.5.2
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
